+++ devices/floppy/swim3.h
@@ -152,7 +152,7 @@ private:
 
     // Interrupt related stuff
     InterruptCtrl* int_ctrl = nullptr;
-    uint32_t       irq_id   = 0;
+    uint64_t       irq_id   = 0;
     uint8_t        irq      = 0;
 };
